feat: implemented delete account modal
